add ergodox compact case
officially abandoning the sixtysexy in favor of a
symmetrical keyboard more closely resembling the
ergodox. This is a new experiment, that will hopefully
yeild more fruit, than the sixtysexy. The key design
choices made here:
64 keys, in 5x6 matrix with 2x2 inner columns
attempt to maintain semblance of ergodox two
primary thumb keys, letting the 4 inner, bottom
row keys be thumb keys.
Still maintain the same form factor as the sixty sexy
altually a bit smaller at 300x130
Put a spot for the teensy in the inner 3 layers, in attempt
to easy soldering pains and remove need for addional bottom
spacer to fit the teensys mini usb port.
Hopefully fully symetrical side to side, stil working on my
autocad skills so may not be the case, but that was teh intent